Move zustand mock
